requirements.txt
Flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.11.1
python-engineio==4.8.2
eventlet==0.33.3
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
pytz==2023.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
filetype==1.2.0